Add missing .edtr extension to the surveys and switch to promises in the process
The surveys differ from the editors in that they are not are a file
present in the users editor directory therefore lack of an extension.

For consistency with other parts in the project an extension is added,
in addition the logic was flatten out replacing the use of callbacks
for promises.

Fix edina/fieldtrip-cobweb-project#89
Fix edina/fieldtrip-cobweb-project#75